+++ src/lib.rs
@@ -1,30 +1,4 @@
-extern crate irc;
-extern crate rusqlite;
-extern crate docopt;
-#[macro_use]
-extern crate serde_derive;
-extern crate itertools;
-extern crate regex;
-#[macro_use]
-extern crate lazy_static;
-#[macro_use]
-extern crate failure;
-extern crate chrono;
-extern crate reqwest;
-extern crate cookie;
-extern crate image;
-extern crate serde_rusqlite;
-extern crate mime;
-extern crate humansize;
-extern crate unicode_segmentation;
-extern crate scraper;
-extern crate toml;
-extern crate directories;
-#[macro_use]
-extern crate log;
-extern crate atty;
-extern crate stderrlog;
-extern crate phf;
+use lazy_static::lazy_static;
 
 pub mod sqlite;
 pub mod http;
@@ -32,6 +6,7 @@ pub mod title;
 pub mod config;
 pub mod message;
 pub mod tld;
+pub mod plugins;
 pub mod buildinfo {
    include!(concat!(env!("OUT_DIR"), "/built.rs"));
 }